+++ poseidon-parameters/Cargo.toml
@@ -8,6 +8,7 @@ license = "MIT OR Apache-2.0"
 keywords = ["no_std"]
 
 [dependencies]
+anyhow = { version = "1", default-features = false }
 ark-ff = { version = "^0.3.0", default-features = false }
 ark-std = { version = "^0.3.0", default-features = false }
 num-integer = { version = "0.1.45", default-features = false }
