+++ poseidon-paramgen/src/lib.rs
@@ -35,9 +35,7 @@ pub mod poseidon_build;
 pub use alpha::Alpha;
 use ark_ff::PrimeField;
 pub use input::InputParameters;
-pub use matrix::{
-    dot_product, mat_mul, Matrix, MatrixOperations, SquareMatrix, SquareMatrixOperations,
-};
+pub use matrix::{dot_product, mat_mul, Matrix, SquareMatrix, SquareMatrixOperations};
 pub use mds::{MdsMatrix, OptimizedMdsMatrices};
 pub use round_constants::{ArcMatrix, OptimizedArcMatrix};
 pub use rounds::RoundNumbers;
