+++ poseidon-paramgen/src/matrix/traits.rs
@@ -1,14 +1,14 @@
 use anyhow::Result;
 
-/// Basic matrix operations all matrices should implement.
-pub trait MatrixOperations<F> {
-    /// Compute matrix transpose
-    fn transpose(&self) -> Self;
-    /// Compute Hadamard (element-wise) product
-    fn hadamard_product(&self, rhs: &Self) -> Result<Self>
-    where
-        Self: Sized;
-}
+// /// Basic matrix operations all matrices should implement.
+// pub trait ExtendedMatrixOperations<F> {
+//     // /// Compute matrix transpose
+//     // fn transpose(&self) -> Self;
+//     /// Compute Hadamard (element-wise) product
+//     fn hadamard_product(&self, rhs: &Self) -> Result<Self>
+//     where
+//         Self: Sized;
+// }
 
 /// Matrix operations that are defined on square matrices.
 pub trait SquareMatrixOperations<F> {
