+++ poseidon-paramgen/src/poseidon_build.rs
@@ -3,7 +3,7 @@ use std::fmt::Display;
 use ark_ff::PrimeField;
 use ark_std::vec::Vec;
 use num::BigUint;
-use poseidon_parameters::BasicMatrixOperations;
+use poseidon_parameters::MatrixOperations;
 
 use crate::{
     Alpha, ArcMatrix, Matrix, MdsMatrix, OptimizedArcMatrix, OptimizedMdsMatrices,
@@ -18,7 +18,7 @@ pub fn compile<F: PrimeField>(
     allow_inverse: bool,
 ) -> String {
     let mut params_code = "use ark_ff::PrimeField;\n
-use poseidon_parameters::{Alpha, ArcMatrix, RoundNumbers, SquareMatrix, Matrix, MdsMatrix, OptimizedArcMatrix, OptimizedMdsMatrices, PoseidonParameters, BasicMatrixOperations};\n\n"
+use poseidon_parameters::{Alpha, ArcMatrix, RoundNumbers, SquareMatrix, Matrix, MdsMatrix, OptimizedArcMatrix, OptimizedMdsMatrices, PoseidonParameters, MatrixOperations};\n\n"
         .to_string();
 
     for t in t_values {
